Refactor: Connection execute function now returns a result object containing either the result or an error.
It's better practise to always return the error rather than throwing it, such that the user can decide what to do with the error. The Result type (Concept from functional programming called Either) ensures that potential errors are always handled by the user before the result can be used. Additionally, the syntax is very clean when used.